pyproject.toml
[tool.poetry]
name = "django-spicy-id"
version = "1.0.0"
description = ""
authors = ["mike wakerly <opensource@hoho.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
tox = "^4.6.3"
setuptools = "^65.7.0"
pytest-django = "^4.5.2"
sphinx = "^6.1.3"
twine = "^4.0.2"
ruff = "^0.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = 'py312'
exclude = [
  ".git",
  "node_modules",
  ".pyenv",
  ".vscode",
  ".tox"
]